+++ src/main/java/kitchenpos/product/application/ProductService.java
@@ -1,6 +1,8 @@
 package kitchenpos.product.application;
 
-import kitchenpos.domain.*;
+import kitchenpos.menus.domain.Menu;
+import kitchenpos.menus.domain.MenuProduct;
+import kitchenpos.menus.domain.MenuRepository;
 import kitchenpos.profanity.domain.PurgomalumClient;
 import kitchenpos.product.domain.Product;
 import kitchenpos.product.domain.ProductRepository;
