+++ src/test/java/kitchenpos/application/OrderServiceTest.java
@@ -1,6 +1,8 @@
 package kitchenpos.application;
 
 import kitchenpos.domain.*;
+import kitchenpos.menus.domain.MenuRepository;
+import kitchenpos.menus.infrastructure.InMemoryMenuRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
