+++ src/test/java/kitchenpos/menus/infrastructure/InMemoryMenuRepository.java
@@ -1,7 +1,7 @@
-package kitchenpos.application;
+package kitchenpos.menus.infrastructure;
 
-import kitchenpos.domain.Menu;
-import kitchenpos.domain.MenuRepository;
+import kitchenpos.menus.domain.Menu;
+import kitchenpos.menus.domain.MenuRepository;
 
 import java.util.*;
 import java.util.stream.Collectors;
