+++ src/test/java/kitchenpos/product/application/ProductServiceTest.java
@@ -1,9 +1,9 @@
 package kitchenpos.product.application;
 
 import kitchenpos.profanity.infrastructure.FakePurgomalumClient;
-import kitchenpos.application.InMemoryMenuRepository;
-import kitchenpos.domain.Menu;
-import kitchenpos.domain.MenuRepository;
+import kitchenpos.menus.infrastructure.InMemoryMenuRepository;
+import kitchenpos.menus.domain.Menu;
+import kitchenpos.menus.domain.MenuRepository;
 import kitchenpos.product.infrastructure.InMemoryProductRepository;
 import kitchenpos.product.domain.Product;
 import kitchenpos.product.domain.ProductRepository;
